Extra Credit Completed
Extra Credit: Added a UISlider to UI which travels through the history
of the currently-being- played game’s flips and displays it to the
user. "The past text" should grey out. Text label never gets blank. And
every time a new flip happens, it “jumps to the present” in the slider.